Make managed-bins feature a default
Working and production builds of the Cicada CLI are not able to see installations of Buildkit (buildctl) on WSL Ubuntu 22.04.2 LTS.

This may potentially affect other distros as well.

Running with the optional managed-bins feature solves the issue.

Though the feature requires a networking dependency that isn't ideal, it should be made default for now.